Removing Popup library, and refactoring the idea
