+++ package.json
@@ -30,8 +30,7 @@
     "react-helmet": "^6.1.0",
     "react-input-slider": "^6.0.0",
     "react-scroll": "^1.8.2",
-    "react-spring": "^9.2.3",
-    "reactjs-popup": "^2.0.4"
+    "react-spring": "^9.2.3"
   },
   "devDependencies": {
     "autoprefixer": "^10.2.6",
