+++ src/components/Advantages/styles.js
@@ -1,0 +1,20 @@
+import styled from "@emotion/styled";
+
+const style = (Component) => styled(Component)`
+
+
+
+
+
+
+.texto{
+font-size: 14px;
+font-weight: 400;
+line-height: 18px;
+letter-spacing: 0em;
+}
+
+
+`;
+
+export default style;
